Implement any/all methods (#49)
These are several methods for `any` and `all` and one method for
`isone`. Methods `any` and `all` are used in `Diagonal`, so they make
other methods efficient.

The tests included in this PR are designed to check the new
methods. But, there is no verification that the new methods are called
rather than the generic methods.

Because the methods for any/all/isone/iszero etc. are interrelated, a
benchmark suite would complement this commit.

Add corner case tests