Make error messages lazy strings (#420)
Based on #419 since `LazyString` and `lazy"..."` are only available in
Julia >= 1.8.

---

The PR changes all error messages to lazy strings to delay string
interpolation and `print`ing of the involved quantities until the error
message is displayed (and hence skip it completely if the error path is
not hit). Using `LazyString`s in error messages generally also helps
with type inference (see e.g.
https://github.com/JuliaLang/julia/pull/58672 and
https://github.com/JuliaLang/julia/pull/54737).

---------

Co-authored-by: Jishnu Bhattacharya <jishnub.github@gmail.com>